# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gooru Serve API
  version: 0.1.0
  description: Initial API contract for the gooru serve web application stack.
servers:
  - url: /api/v1
paths:
...
  /files:
    get:
      summary: List tracked files for browser grids.
      security:
        - sessionAuth: []
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: page_token
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [added, name, modified, size, kind]
            default: added
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: include_facets
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Page of tracked files.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      summary: Untrack or physically delete a selected set of files.
      description: Accepts the same explicit-ID or query-with-exclusions selector used by bulk UI actions. Strict physical deletion is accepted only when every selected file is inside configured upload targets. After explicit confirmation, delete_or_untrack deletes managed files while only untracking files outside configured upload targets.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRemovalRequest"
      responses:
        "200":
          description: Selected file locations were removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRemovalResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Physical deletion was requested for a selection containing files outside configured upload targets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /files/{id}:
    get:
      summary: Get one tracked file by opaque file ID.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracked file details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Untrack a file location or delete managed uploaded content.
      description: Untrack removes only the selected database location. Delete also removes the underlying file, but is accepted only for paths inside configured upload targets.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  enum: [untrack, delete]
                  default: untrack
      responses:
        "200":
          description: File location was untracked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteFileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Physical deletion was requested for a file outside configured upload targets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    CSRF:
      name: X-Gooru-CSRF
      in: header
      required: true
      schema:
        type: string
      description: CSRF token returned by /auth/login or /auth/me. Required for cookie-authenticated mutating requests.
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: gooru_session
  responses:
    Unauthorized:
      description: Missing or invalid session cookie.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Request is forbidden, including failed CSRF validation or insufficient privileges.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
...
    FileRemovalSelector:
      type: object
      properties:
        file_ids:
          type: array
          items:
            type: string
        query:
          type: string
        selection_id:
          type: string
          description: Immutable selection snapshot ID. Mutually exclusive with file_ids and query.
        include_file_ids:
          type: array
          items:
            type: string
          description: Explicit additions to a selection snapshot. Valid only with selection_id.
        exclude_file_ids:
          type: array
          items:
            type: string
          description: Explicit exclusions from a query or selection snapshot.
    FileRemovalRequest:
      allOf:
        - $ref: "#/components/schemas/FileRemovalSelector"
        - type: object
          required: [mode]
          properties:
            mode:
              type: string
              enum: [untrack, delete, delete_or_untrack]
    FileRemovalResponse:
      type: object
      required: [mode, selector, removed_locations]
      properties:
        mode:
          type: string
          enum: [untrack, delete, delete_or_untrack]
        selector:
          $ref: "#/components/schemas/FileRemovalSelector"
        removed_locations:
          type: integer
          minimum: 0
...
    FileListResponse:
      type: object
      required: [files, total_count, library_count]
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/File"
        next_page_token:
          type: string
        previous_page_token:
          type: string
        total_count:
          type: integer
          description: Total number of files matching the current query.
        library_count:
          type: integer
          description: Total tracked files in the library.
        facets:
          $ref: "#/components/schemas/Facets"
    Facets:
      type: object
      properties:
        kind:
          type: array
          items:
            $ref: "#/components/schemas/FacetValue"
    FacetValue:
      type: object
      required: [value, count]
      properties:
        value:
          type: string
        count:
          type: integer
    File:
      type: object
      required:
        - id
        - content_id
        - name
        - safe_display_path
        - size
        - added_at
        - modified_time
        - media_type
        - media_kind
        - viewer_support
        - metadata
        - tags
        - media_urls
        - can_delete
      properties:
        id:
          type: string
          description: Opaque stable file-location identifier.
        content_id:
          type: string
          description: Content fingerprint for the tracked file.
        name:
          type: string
        path:
          type: string
          description: Absolute filesystem path. Omitted unless server path exposure is enabled.
        safe_display_path:
          type: string
          description: Display-only non-absolute path hint, such as the last directory plus filename.
        size:
          type: integer
          format: int64
        added_at:
          type: string
          format: date-time
        modified_time:
          type: string
          format: date-time
        media_type:
          type: string
        media_kind:
          type: string
          enum: [photo, video, gif, audio, comic, other]
        viewer_support:
          type: string
          enum: [supported, unsupported_media_type]
          description: Backend-owned indication of whether the built-in viewer supports this media type.
        metadata:
          $ref: "#/components/schemas/MediaMetadata"
        tags:
          type: array
          items:
            type: string
        media_urls:
          $ref: "#/components/schemas/MediaURLs"
        can_delete:
          type: boolean
          description: True when the file path is inside a configured upload target and may be physically deleted through the API.
    MediaMetadata:
      type: object
      description: Optional media metadata. Fields are omitted when unavailable or extraction fails.
      properties:
        image_width:
          type: integer
        image_height:
          type: integer
        video_width:
          type: integer
        video_height:
          type: integer
        video_duration:
          type: number
          format: double
        audio_duration:
          type: number
          format: double
        frame_count:
          type: integer
        page_count:
          type: integer
          minimum: 1
          description: Number of displayable pages for paged media such as CBZ comics.
    MediaURLs:
      type: object
      required: [thumbnail, preview, content, download]
      properties:
        thumbnail:
          type: string
        preview:
          type: string
        content:
          type: string
        download:
          type: string
...
    DeleteFileResponse:
      type: object
      required: [mode, removed_locations]
      properties:
        mode:
          type: string
          enum: [untrack, delete]
        removed_locations:
          type: integer
...